# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-Cloud Terrarium (mc-terrarium) aims to provide an environment to enrich multi-cloud infrastructure.",
        "title": "Multi-Cloud Terrarium REST API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/cloud-barista/mc-terrarium/issues/new"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0.1.2"
    },
    "host": "localhost:8055",
    "basePath": "/terrarium",
    "paths": {
...
        "/tr/{trId}/vpn/gcp-aws": {
            "get": {
                "description": "Get resource info to configure GCP to AWS VPN tunnels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[VPN] GCP to AWS VPN tunnel configuration (PoC - Not officially supported)"
                ],
                "summary": "Get resource info to configure GCP to AWS VPN tunnels",
                "parameters": [
                    {
                        "type": "string",
                        "default": "tr01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "refined",
                        "description": "Resource info by detail (refined, raw)",
                        "name": "detail",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Refresh the state before getting the info",
                        "name": "refresh",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create network resources for VPN tunnel in GCP and AWS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[VPN] GCP to AWS VPN tunnel configuration (PoC - Not officially supported)"
                ],
                "summary": "Create network resources for VPN tunnel in GCP and AWS",
                "parameters": [
                    {
                        "type": "string",
                        "default": "tr01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Destroy network resources that were used to configure GCP as an AWS VPN tunnel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[VPN] GCP to AWS VPN tunnel configuration (PoC - Not officially supported)"
                ],
                "summary": "Destroy network resources that were used to configure GCP as an AWS VPN tunnel",
                "parameters": [
                    {
                        "type": "string",
                        "default": "tr01",
                        "description": "Terrarium ID",
                        "name": "trId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Custom request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Credential holder (profile) name",
                        "name": "X-Credential-Holder",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Response": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "Any details"
                },
                "list": {
                    "type": "array",
                    "items": {}
                },
                "message": {
                    "type": "string",
                    "example": "Any message"
                },
                "object": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}